+++ src/components/Main.js
@@ -1,12 +1,12 @@
 import React from "react";
 import styled from "styled-components";
-import SlideView from "./shared/SlideView/SlideView";
+import SlideList from "./shared/SlideView/SlideView";
 
 function Main() {
   return (
     <MainContainer>
-      <SlideView />
-      <SlideView />
+      <SlideList />
+      <SlideList />
     </MainContainer>
   );
 }
